refactor: apply /simplify findings (Hermes worker blocking, test dedup)
- Hermes sync_turn: the prior-thread join and transcript formatting now
  run inside the newly-spawned background thread instead of on the
  caller. The caller is Hermes's own shared single-worker executor, so
  a synchronous join(timeout=5.0) there could block it for up to 5s on
  every turn — exactly the backlog the background thread exists to
  avoid. Bounded thread growth (join-before-respawn) is preserved, just
  relocated off the shared worker.
- Extract the duplicated agent-session test-row builder (sessions-filters
  and session-detail-curation tests) into a shared apps/server/src/test
  fixture.
- Consolidation runner tests that don't exercise purgeEmpty now pass a
  minimal stub instead of constructing a full AgentSessionsService.
