+++ apps/server/src/consolidation/orphan-promotion.test.ts
@@ -4,7 +4,6 @@ import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 import { createRepositories } from '../db/repositories/index.js';
 import { consolidationOps } from '../db/schema/consolidation.js';
 import { memoryRelations } from '../db/schema/memory-relations.js';
-import { AgentSessionsService } from '../services/agent-sessions.js';
 import { MemoryService } from '../services/memory.js';
 import { ProjectsService } from '../services/projects.js';
 import { RelationsService } from '../services/relations.js';
@@ -43,7 +42,7 @@ beforeEach(() => {
     repos: createRepositories(db.handle.db),
     tx: db.handle.db,
     relations,
-    agentSessions: new AgentSessionsService(createRepositories(db.handle.db), db.handle.db),
+    agentSessions: { purgeEmpty: () => ({ deletedIds: [] }) },
     orphanDeadlineMs,
   });
 });
