+++ apps/server/src/dashboard/session-detail-curation.test.ts
@@ -4,24 +4,15 @@ import { Hono, type Context, type Next } from 'hono';
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 
 import { createRepositories } from '../db/repositories/index.js';
-import { agentSessions, type NewAgentSession } from '../db/schema/agent-sessions.js';
+import { agentSessions } from '../db/schema/agent-sessions.js';
 import { tokens } from '../db/schema/tokens.js';
 import { AgentSessionsService } from '../services/agent-sessions.js';
 import { SessionsService } from '../services/sessions.js';
 import { TokensService } from '../services/tokens.js';
-import { createTestDb, type TestDb } from '../test/db.js';
+import { agentSessionRow as row, createTestDb, type TestDb } from '../test/index.js';
 
 import { createSessionsRouter } from './sessions.js';
 import type { ResolvedSession } from './types.js';
-
-function row(overrides: Partial<NewAgentSession> & { id: string }): NewAgentSession {
-  return {
-    tokenId: 'tk1',
-    agent: 'claude-code',
-    startedAt: new Date(1_000),
-    ...overrides,
-  };
-}
 
 describe('session detail curation-state rendering', () => {
   let t: TestDb;
